Refactored modelcontroller
Modelcontroller._displayed_machines stored tuples of information that
was not automatically updated. Changed so that it know stores
ProvenanceClient instances and there is a new _refresh_hosts methods
that is called when calling the displayed_hosts getter. Also, when the
displayed_hosts getter is called it doesn't return the list it returns
the original tuples of information